Added support of different edge version for checkpoint rule node
