+++ application/src/main/java/org/thingsboard/server/service/edge/rpc/constructor/RuleChainMsgConstructor.java
@@ -15,44 +15,31 @@
  */
 package org.thingsboard.server.service.edge.rpc.constructor;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
+import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.thingsboard.common.util.JacksonUtil;
-import org.thingsboard.rule.engine.flow.TbRuleChainInputNode;
-import org.thingsboard.rule.engine.flow.TbRuleChainInputNodeConfiguration;
-import org.thingsboard.rule.engine.flow.TbRuleChainOutputNode;
 import org.thingsboard.server.common.data.id.RuleChainId;
-import org.thingsboard.server.common.data.rule.NodeConnectionInfo;
+import org.thingsboard.server.common.data.id.TenantId;
 import org.thingsboard.server.common.data.rule.RuleChain;
-import org.thingsboard.server.common.data.rule.RuleChainConnectionInfo;
 import org.thingsboard.server.common.data.rule.RuleChainMetaData;
-import org.thingsboard.server.common.data.rule.RuleNode;
+import org.thingsboard.server.dao.queue.QueueService;
 import org.thingsboard.server.gen.edge.v1.EdgeVersion;
-import org.thingsboard.server.gen.edge.v1.NodeConnectionInfoProto;
-import org.thingsboard.server.gen.edge.v1.RuleChainConnectionInfoProto;
 import org.thingsboard.server.gen.edge.v1.RuleChainMetadataUpdateMsg;
 import org.thingsboard.server.gen.edge.v1.RuleChainUpdateMsg;
-import org.thingsboard.server.gen.edge.v1.RuleNodeProto;
 import org.thingsboard.server.gen.edge.v1.UpdateMsgType;
 import org.thingsboard.server.queue.util.TbCoreComponent;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.NavigableSet;
-import java.util.TreeSet;
-import java.util.UUID;
-import java.util.stream.Collectors;
+import org.thingsboard.server.service.edge.rpc.constructor.rule.RuleChainMetadataConstructor;
+import org.thingsboard.server.service.edge.rpc.constructor.rule.RuleChainMetadataConstructorFactory;
 
 @Component
 @Slf4j
 @TbCoreComponent
+@AllArgsConstructor
 public class RuleChainMsgConstructor {
 
-    private static final String RULE_CHAIN_INPUT_NODE = TbRuleChainInputNode.class.getName();
-    private static final String TB_RULE_CHAIN_OUTPUT_NODE = TbRuleChainOutputNode.class.getName();
+    private final QueueService queueService;
 
     public RuleChainUpdateMsg constructRuleChainUpdatedMsg(RuleChainId edgeRootRuleChainId, UpdateMsgType msgType, RuleChain ruleChain) {
         RuleChainUpdateMsg.Builder builder = RuleChainUpdateMsg.newBuilder()
@@ -70,229 +57,13 @@ public class RuleChainMsgConstructor {
         return builder.build();
     }
 
-    public RuleChainMetadataUpdateMsg constructRuleChainMetadataUpdatedMsg(UpdateMsgType msgType,
+    public RuleChainMetadataUpdateMsg constructRuleChainMetadataUpdatedMsg(TenantId tenantId,
+                                                                           UpdateMsgType msgType,
                                                                            RuleChainMetaData ruleChainMetaData,
                                                                            EdgeVersion edgeVersion) {
-        try {
-            RuleChainMetadataUpdateMsg.Builder builder = RuleChainMetadataUpdateMsg.newBuilder();
-            switch (edgeVersion) {
-                case V_3_3_0:
-                    constructRuleChainMetadataUpdatedMsg_V_3_3_0(builder, ruleChainMetaData);
-                    break;
-                case V_3_3_3:
-                default:
-                    constructRuleChainMetadataUpdatedMsg_V_3_3_3(builder, ruleChainMetaData);
-                    break;
-            }
-            builder.setMsgType(msgType);
-            return builder.build();
-        } catch (JsonProcessingException ex) {
-            log.error("Can't construct RuleChainMetadataUpdateMsg", ex);
-        }
-        return null;
-    }
-
-    private void constructRuleChainMetadataUpdatedMsg_V_3_3_3(RuleChainMetadataUpdateMsg.Builder builder,
-                                                                                           RuleChainMetaData ruleChainMetaData) throws JsonProcessingException {
-        builder.setRuleChainIdMSB(ruleChainMetaData.getRuleChainId().getId().getMostSignificantBits())
-                .setRuleChainIdLSB(ruleChainMetaData.getRuleChainId().getId().getLeastSignificantBits())
-                .addAllNodes(constructNodes(ruleChainMetaData.getNodes()))
-                .addAllConnections(constructConnections(ruleChainMetaData.getConnections()))
-                .addAllRuleChainConnections(constructRuleChainConnections(ruleChainMetaData.getRuleChainConnections(), new TreeSet<>()));
-        if (ruleChainMetaData.getFirstNodeIndex() != null) {
-            builder.setFirstNodeIndex(ruleChainMetaData.getFirstNodeIndex());
-        } else {
-            builder.setFirstNodeIndex(-1);
-        }
-    }
-
-    private void constructRuleChainMetadataUpdatedMsg_V_3_3_0(RuleChainMetadataUpdateMsg.Builder builder,
-                                                                                           RuleChainMetaData ruleChainMetaData) throws JsonProcessingException {
-        List<RuleNode> supportedNodes = filterNodes_V_3_3_0(ruleChainMetaData.getNodes());
-        NavigableSet<Integer> removedNodeIndexes = getRemovedNodeIndexes(ruleChainMetaData.getNodes(), ruleChainMetaData.getConnections());
-        List<NodeConnectionInfo> connections = filterConnections_V_3_3_0(ruleChainMetaData.getNodes(), ruleChainMetaData.getConnections(), removedNodeIndexes);
-
-        List<RuleChainConnectionInfo> ruleChainConnections = new ArrayList<>();
-        if (ruleChainMetaData.getRuleChainConnections() != null) {
-            ruleChainConnections.addAll(ruleChainMetaData.getRuleChainConnections());
-        }
-        ruleChainConnections.addAll(addRuleChainConnections_V_3_3_0(ruleChainMetaData.getNodes(), ruleChainMetaData.getConnections()));
-        builder.setRuleChainIdMSB(ruleChainMetaData.getRuleChainId().getId().getMostSignificantBits())
-                .setRuleChainIdLSB(ruleChainMetaData.getRuleChainId().getId().getLeastSignificantBits())
-                .addAllNodes(constructNodes(supportedNodes))
-                .addAllConnections(constructConnections(connections))
-                .addAllRuleChainConnections(constructRuleChainConnections(ruleChainConnections, removedNodeIndexes));
-        if (ruleChainMetaData.getFirstNodeIndex() != null) {
-            Integer firstNodeIndex = ruleChainMetaData.getFirstNodeIndex();
-            // decrease index because of removed nodes
-            for (Integer removedIndex : removedNodeIndexes) {
-                if (firstNodeIndex > removedIndex) {
-                    firstNodeIndex = firstNodeIndex - 1;
-                }
-            }
-            builder.setFirstNodeIndex(firstNodeIndex);
-        } else {
-            builder.setFirstNodeIndex(-1);
-        }
-    }
-
-    private List<NodeConnectionInfo> filterConnections_V_3_3_0(List<RuleNode> nodes, List<NodeConnectionInfo> connections, NavigableSet<Integer> removedNodeIndexes) {
-        List<NodeConnectionInfo> result = new ArrayList<>();
-        if (connections != null) {
-            result = connections.stream().filter(conn -> {
-                for (int i = 0; i < nodes.size(); i++) {
-                    RuleNode node = nodes.get(i);
-                    if (node.getType().equalsIgnoreCase(RULE_CHAIN_INPUT_NODE)
-                            || node.getType().equalsIgnoreCase(TB_RULE_CHAIN_OUTPUT_NODE)) {
-                        if (conn.getFromIndex() == i || conn.getToIndex() == i) {
-                            return false;
-                        }
-                    }
-                }
-                return true;
-            }).map(conn -> {
-                NodeConnectionInfo newConn = new NodeConnectionInfo();
-                newConn.setFromIndex(conn.getFromIndex());
-                newConn.setToIndex(conn.getToIndex());
-                newConn.setType(conn.getType());
-                return newConn;
-            }).collect(Collectors.toList());
-        }
-
-        // decrease index because of removed nodes
-        for (Integer removedIndex : removedNodeIndexes) {
-            for (NodeConnectionInfo newConn : result) {
-                if (newConn.getToIndex() > removedIndex) {
-                    newConn.setToIndex(newConn.getToIndex() - 1);
-                }
-                if (newConn.getFromIndex() > removedIndex) {
-                    newConn.setFromIndex(newConn.getFromIndex() - 1);
-                }
-            }
-        }
-
-        return result;
-    }
-
-    private NavigableSet<Integer> getRemovedNodeIndexes(List<RuleNode> nodes, List<NodeConnectionInfo> connections) {
-        TreeSet<Integer> removedIndexes = new TreeSet<>();
-        for (NodeConnectionInfo connection : connections) {
-            for (int i = 0; i < nodes.size(); i++) {
-                RuleNode node = nodes.get(i);
-                if (node.getType().equalsIgnoreCase(RULE_CHAIN_INPUT_NODE)
-                        || node.getType().equalsIgnoreCase(TB_RULE_CHAIN_OUTPUT_NODE)) {
-                    if (connection.getFromIndex() == i || connection.getToIndex() == i) {
-                        removedIndexes.add(i);
-                    }
-                }
-            }
-        }
-        return removedIndexes.descendingSet();
-    }
-
-    private List<NodeConnectionInfoProto> constructConnections(List<NodeConnectionInfo> connections) {
-        List<NodeConnectionInfoProto> result = new ArrayList<>();
-        if (connections != null && !connections.isEmpty()) {
-            for (NodeConnectionInfo connection : connections) {
-                result.add(constructConnection(connection));
-            }
-        }
-        return result;
-    }
-
-    private NodeConnectionInfoProto constructConnection(NodeConnectionInfo connection) {
-        return NodeConnectionInfoProto.newBuilder()
-                .setFromIndex(connection.getFromIndex())
-                .setToIndex(connection.getToIndex())
-                .setType(connection.getType())
-                .build();
-    }
-
-    private List<RuleNode> filterNodes_V_3_3_0(List<RuleNode> nodes) {
-        List<RuleNode> result = new ArrayList<>();
-        for (RuleNode node : nodes) {
-            if (RULE_CHAIN_INPUT_NODE.equals(node.getType())
-                    || TB_RULE_CHAIN_OUTPUT_NODE.equals(node.getType())) {
-                log.trace("Skipping not supported rule node {}", node);
-            } else {
-                result.add(node);
-            }
-        }
-        return result;
-    }
-
-    private List<RuleNodeProto> constructNodes(List<RuleNode> nodes) throws JsonProcessingException {
-        List<RuleNodeProto> result = new ArrayList<>();
-        if (nodes != null && !nodes.isEmpty()) {
-            for (RuleNode node : nodes) {
-                result.add(constructNode(node));
-            }
-        }
-        return result;
-    }
-
-    private List<RuleChainConnectionInfo> addRuleChainConnections_V_3_3_0(List<RuleNode> nodes, List<NodeConnectionInfo> connections) throws JsonProcessingException {
-        List<RuleChainConnectionInfo> result = new ArrayList<>();
-        for (int i = 0; i < nodes.size(); i++) {
-            RuleNode node = nodes.get(i);
-            if (node.getType().equalsIgnoreCase(RULE_CHAIN_INPUT_NODE)) {
-                for (NodeConnectionInfo connection : connections) {
-                    if (connection.getToIndex() == i) {
-                        RuleChainConnectionInfo e = new RuleChainConnectionInfo();
-                        e.setFromIndex(connection.getFromIndex());
-                        TbRuleChainInputNodeConfiguration configuration = JacksonUtil.treeToValue(node.getConfiguration(), TbRuleChainInputNodeConfiguration.class);
-                        e.setTargetRuleChainId(new RuleChainId(UUID.fromString(configuration.getRuleChainId())));
-                        e.setAdditionalInfo(node.getAdditionalInfo());
-                        e.setType(connection.getType());
-                        result.add(e);
-                    }
-                }
-            }
-        }
-        return result;
-    }
-
-    private List<RuleChainConnectionInfoProto> constructRuleChainConnections(List<RuleChainConnectionInfo> ruleChainConnections, NavigableSet<Integer> removedNodeIndexes) throws JsonProcessingException {
-        List<RuleChainConnectionInfoProto> result = new ArrayList<>();
-        if (ruleChainConnections != null && !ruleChainConnections.isEmpty()) {
-            for (RuleChainConnectionInfo ruleChainConnectionInfo : ruleChainConnections) {
-                result.add(constructRuleChainConnection(ruleChainConnectionInfo, removedNodeIndexes));
-            }
-        }
-        return result;
-    }
-
-    private RuleChainConnectionInfoProto constructRuleChainConnection(RuleChainConnectionInfo ruleChainConnectionInfo, NavigableSet<Integer> removedNodeIndexes) throws JsonProcessingException {
-        int fromIndex = ruleChainConnectionInfo.getFromIndex();
-        // decrease index because of removed nodes
-        for (Integer removedIndex : removedNodeIndexes) {
-            if (fromIndex > removedIndex) {
-                fromIndex = fromIndex - 1;
-            }
-        }
-        ObjectNode additionalInfo = (ObjectNode) ruleChainConnectionInfo.getAdditionalInfo();
-        if (additionalInfo.get("ruleChainNodeId") == null) {
-            additionalInfo.put("ruleChainNodeId", "rule-chain-node-UNDEFINED");
-        }
-        return RuleChainConnectionInfoProto.newBuilder()
-                .setFromIndex(fromIndex)
-                .setTargetRuleChainIdMSB(ruleChainConnectionInfo.getTargetRuleChainId().getId().getMostSignificantBits())
-                .setTargetRuleChainIdLSB(ruleChainConnectionInfo.getTargetRuleChainId().getId().getLeastSignificantBits())
-                .setType(ruleChainConnectionInfo.getType())
-                .setAdditionalInfo(JacksonUtil.OBJECT_MAPPER.writeValueAsString(additionalInfo))
-                .build();
-    }
-
-    private RuleNodeProto constructNode(RuleNode node) throws JsonProcessingException {
-        return RuleNodeProto.newBuilder()
-                .setIdMSB(node.getId().getId().getMostSignificantBits())
-                .setIdLSB(node.getId().getId().getLeastSignificantBits())
-                .setType(node.getType())
-                .setName(node.getName())
-                .setDebugMode(node.isDebugMode())
-                .setConfiguration(JacksonUtil.OBJECT_MAPPER.writeValueAsString(node.getConfiguration()))
-                .setAdditionalInfo(JacksonUtil.OBJECT_MAPPER.writeValueAsString(node.getAdditionalInfo()))
-                .build();
+        RuleChainMetadataConstructor ruleChainMetadataConstructor
+                = RuleChainMetadataConstructorFactory.getByEdgeVersion(edgeVersion, queueService);
+        return ruleChainMetadataConstructor.constructRuleChainMetadataUpdatedMsg(tenantId, msgType, ruleChainMetaData);
     }
 
     public RuleChainUpdateMsg constructRuleChainDeleteMsg(RuleChainId ruleChainId) {
@@ -301,5 +72,4 @@ public class RuleChainMsgConstructor {
                 .setIdMSB(ruleChainId.getId().getMostSignificantBits())
                 .setIdLSB(ruleChainId.getId().getLeastSignificantBits()).build();
     }
-
 }
